Store seed stock cost in öre (cents) to avoid floating-point rounding
Backend stays Int (öre). Frontend converts:
- Input: user types SEK (e.g. 0.25), sends öre (25) to API
- Display: API returns öre (25), shows as "0.25 kr"
- Edit prefill: converts öre back to SEK for input fields

+++ backend/src/main/kotlin/app/verdant/dto/SeedInventoryDtos.kt
@@ -14,7 +14,7 @@ data class SeedInventoryResponse(
     val quantity: Int,
     val collectionDate: LocalDate?,
     val expirationDate: LocalDate?,
-    val costPerUnitSek: Double?,
+    val costPerUnitSek: Int?,
     val unitType: String?,
     val seasonId: Long?,
     val speciesProviderId: Long?,
@@ -30,7 +30,7 @@ data class CreateSeedInventoryRequest(
     val collectionDate: LocalDate? = null,
     val expirationDate: LocalDate? = null,
     @field:Min(0)
-    val costPerUnitSek: Double? = null,
+    val costPerUnitSek: Int? = null,
     @field:NotBlank @field:Size(max = 255)
     val unitType: String,
     val seasonId: Long? = null,
@@ -43,7 +43,7 @@ data class UpdateSeedInventoryRequest(
     val collectionDate: LocalDate? = null,
     val expirationDate: LocalDate? = null,
     @field:Min(0)
-    val costPerUnitSek: Double? = null,
+    val costPerUnitSek: Int? = null,
     @field:Size(max = 255)
     val unitType: String? = null,
     val seasonId: Long? = null,
